+++ cli/src/qod_cli/main.py
@@ -91,11 +91,8 @@ from .commands import sql_cmd  # noqa: E402
 
 app.command("sql")(sql_cmd.sql)
 
-from .commands import demo, start, stop  # noqa: E402
+from .commands import start, stop  # noqa: E402
 
-app.command(
-    "demo", context_settings={"allow_extra_args": True, "ignore_unknown_options": True}
-)(demo.demo)
 app.command(
     "start", context_settings={"allow_extra_args": True, "ignore_unknown_options": True}
 )(start.start)